+++ assets/script/ComTest.js
@@ -1,9 +1,0 @@
-// import { h } from "vue";
-
-export default {
-	methods: {
-		doSomething() {
-			console.log("Yes");
-		}
-	}
-};