+++ route.js
@@ -5,7 +5,7 @@ import "./routes/fonts.js";
 import "./routes/dictionary.js";
 import "./routes/definition.js";
 
-// import "./routes/sponsor.js";
+// import "./routes/donate.js";
 
 import "./routes/privacy-policy.js";
 import "./routes/terms-of-service.js";
@@ -16,3 +16,5 @@ import "./routes/cookie-policy.js";
 
 import "./routes/api.js";
 // import "./routes/office.js";
+
+import "./routes/template.js";