+++ routes/home.js
@@ -5,10 +5,10 @@ const routes = app.routes("/", "page");
 
 routes.register({ name: "home", text: "Home" }, function(req, res) {
 	res.render("home", {
-		title: "Myanmar dictionary",
+		title: "online Myanmar dictionary",
 		keywords: "Myanmar, dictionary, grammar, font, definition, Burmese, online",
 		description:
-			"A comprehensive Myanmar dictionary online, grammar, and fonts at MyOrdbok",
+			"A comprehensive online Myanmar dictionary, grammar, and fonts at MyOrdbok",
 		pageClass: "home"
 	});
 });